Removed IO_raw as there didn't seem to be a need and moved to IO_numpy instead.
